Базовая прогрессия организма и РНА (v0.1)
Реализован цикл адаптации: 24000 активных тиков копят активность, и только
при закрытии цикла она превращается в рост — всплеск спама и ровный час
упираются в один и тот же потолок.

Физика: четыре параметра (Сила, Выносливость, Стойкость, Восстановление)
в отдельном аттачменте, старт 40.0, убывающая отдача по текущему значению.
Источники устойчивы к абузу: движение не считается в транспорте и при
телепорте, добыча взвешена по твёрдости, ближний бой закрывает не больше
четверти нормы Силы, урон от себя и повтор одного источника в коротком окне
не идут в Стойкость, а Восстановление требует настоящего эпизода нагрузки —
едой и сном его не купить.

РНА: параметры переведены в дробные — при int прирост 0.40 за цикл
округлялся в ноль и рост не работал бы вовсе. Практика растёт только от
успешных осмысленных применений, с затуханием повтора примитива 8/16.
Пропускная тренируется дорогими кастами, Устойчивость — работой в верхней
полосе нагрузки, Связность копится скрыто и требует широты, объёма и
отсутствия компенсации за цикл.

Все балансные числа в конфиге, добавлена команда /riftborne power adaptation
с принудительным закрытием цикла для проверки.

+++ src/main/java/com/pr1tcha/riftborne/rna/power/client/PowerHud.java
@@ -95,7 +95,7 @@ public final class PowerHud {
                 critical ? COLOR_WEAR_CRIT : COLOR_WEAR);
 
         // Throughput line: past this a cast overloads the nodes even when it is admissible.
-        int throughputX = x + scale(profile.throughput());
+        int throughputX = x + scale((float) profile.throughput());
         graphics.fill(throughputX, y - 2, throughputX + 1, y + BAR_HEIGHT + 2, COLOR_THROUGHPUT);
 
         // Projection of the armed primitive.
